Rename import/export functions and files
